+++ src/archive/waitscreen.bak.ps1
@@ -1,7 +1,0 @@
-function Write-WaitingScreen {
-  Clear-Host; Write-HeaderLabel
-  Write-Host "`nActivation script for QuickBooks POS." -ForegroundColor White
-  Write-Host "`nPlease ensure that the QuickBooks software is completely"
-  Write-Host "installed on your system. Activation will proceed to begin"
-  Write-Host "after you choose to continue.`n"
-}